Add loading indicator with HTTP interceptor
A loading component and a loading interceptor for HTTP requests were added to improve the user experience during network activity. The loading interceptor uses the new `LoadingService` to toggle a loading indicator on and off for each HTTP request, but can be skipped via `SkipLoading` context. The CoursesService has also been updated to use this bypass option.