Add messages service integration to components
The messages service has been integrated into the necessary components, which allows displaying informative messages to the user. Added a new function to clear the message on user action. Modified the templates to reflect these changes. The home component has been updated to show an error message when there is a problem loading courses.