Add authentication handling to the application
The changes include the implementation of a login and logout mechanic, along with the conditional displaying of these options within the application. The AuthService was significantly expanded to handle user state and interact with the implemented back-end api for login. Additionally, form control was added to the login page to handle user input, and a new guard, isUserAuthenticatedGuard, was created to check user authentication status on certain routes.